Fix some bugs, but not all
Namely, when a Jsonnet buffer is NOT visiable when :TankaOff is run, the
next time you open it, the statusline is still there. This version of
the code tries to use local statuslines to be nice, but I think it's
going to be too much work. It seems like Vim is keeping some state
internally. I really don't want to add an autocmd that is active even
when the rest of the plugin is disabled. It would probably work, but
meh.